Extract langfuse tracing setup for google-adk to more general location
